Change combos alignment since impossible combos are now hidden
For example, a lvl 20 476 hp cog can be killed with 4 bday cakes,
and there are no throw combos that will kill with fewer toons.
Looking across the columns, this variance of the number of combo
rows displayed caused them to be misaligned. I.e a 4 toon combo
shown in one place had a different y-alignment than a combo
in another place. This change makes the alignment consistent by
no longer centering the combos in a cell.